backend/requirements.txt
fastapi
uvicorn
python-keycloak
requests